Villager: Assign to abode or town at creation
The abode position is compared to each door of abodes in 10x10 grids to
assign a villager to an abode.
The town is selected by the abodes in this grid.
If no abode with space is close enough, a random abode in town is selected.
Villager is assigned to town as homeless if no abodes exist.
If no town is found, the villagers should go into a list of villagers
without towns.

+++ src/ECS/Archetypes/TownArchetype.cpp
@@ -25,7 +25,7 @@ entt::entity TownArchetype::Create(int id, const glm::vec3& position, [[maybe_un
 
 	// const auto& info = Game::instance()->GetInfoConstants().town;
 
-	registry.Assign<Town>(entity, id);
+	registry.Assign<Town>(entity, static_cast<uint32_t>(id));
 	registry.Assign<Transform>(entity, position);
 	registry.Assign<Tribe>(entity, tribe);
 	registry.Assign<Transform>(entity, position, glm::mat3(1.0f), glm::vec3(1.0f));
